+++ src/main/java/db/dao/FlightDAO.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 
 public interface FlightDAO {
     long create(Flight flight);
-    Optional<Flight> get(int id);
+    Optional<Flight> get(long id);
     void update(Flight flight);
     void remove(Flight flight);
     List<Flight> getAll();
