+++ src/main/java/db/dao/UserDAO.java
@@ -7,7 +7,7 @@ import java.util.Optional;
 
 public interface UserDAO {
     long create(User user);
-    Optional<User> get(int id);
+    Optional<User> get(long id);
     Optional<User> get(String email);
     void update(User user);
     void remove(User user);
