+++ src/main/java/db/service/UserService.java
@@ -41,7 +41,7 @@ public class UserService implements UserDAO {
 
     @Override
     @SneakyThrows
-    public Optional<User> get(int id) {
+    public Optional<User> get(long id) {
         String sql = "SELECT name, email, password_hash, registration_date FROM Account WHERE id = ?";
 
         try(Connection connection = DataSource.getConnection();
